pyproject.toml
[tool.poetry]
name = "spec_projector"
version = "0.1.0"
description = ""
authors = ["Junte Ltd"]

[tool.poetry.dependencies]
python = "^3.8"
python-decouple = "*"
psycopg2-binary = "*"
django = "3.*"
django-admin-autocomplete-filter = "*"
django-health-check = "*"
django-cors-headers = "*"
django_extensions = "*"
django-split-settings = "*"
django-model-utils = "*"
django-filter = "*"
djangorestframework = "*"
drf-yasg = "*"
jnt-django-bitfield = "*"
jnt-django-admin-tools = "*"
social-auth-app-django = "^3.*"
social-auth-core = "^3.*"
graphene-django = "^2.*"
uWSGI = "*"
python-gitlab = "*"
PyGithub = "*"

[tool.poetry.dev-dependencies]
mypy = "*"
polint = "*"
dennis = "*"
xenon = "*"
safety = "*"
pytest = "*"
pytest-cov = "*"
pytest-django = "*"
pytest-xdist = "*"
httpretty = "*"
factory-boy = "*"
flake8 = "*"
flake8-pytest = "*"
flake8-django = "*"
flake8-pytest-style = "*"
wemake-python-styleguide = "*"
black = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 79
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
(
  /(
    | \.git          # root of the project
    | \.mypy_cache
    | \.pytest_cache
  )/
)
'''
